refactor with functioning cover paths
